+++ src/classes/BloodConfig.ts
