+++ src/classes/BrushControls.ts
